==> Processing/1_EES_database.Rmd <==
---
title: "1_Gather_data"
author: "Arturo Bertero"
date: "2024-03-19"
output: html_document
---

# Libraries

```{r}
#packages
library("pacman")
p_load(tidyverse, here, sjlabelled, stringr, glue, janitor, haven, stargazer, 
       ltm, skimr, readxl, naniar, labelled, conflicted, countrycode, visdat)

#remove scientific notation
options(scipen=999)

#conflicts
conflict_prefer("select", "dplyr")
conflict_prefer("filter", "dplyr")
```

# Input

```{r}
#EES data

#Combined data from 1989 to 2004 
d_1989_2004 = read_dta(here("Input", "data", "trendEES.dta")) %>% 
  clean_names()

#Original data 2009
d_2009 = read_dta(here("Input", "data", "2009.dta")) %>% 
  clean_names()
  
  ##ptvs 
  #table(d_2009$q39_p1) #first
  #table(d_2009$q39_p15) #last
  
#Original data 2014 
d_2014 = read_dta(here("Input", "data", "2014.dta")) %>% 
  clean_names()
  
  ##ptvs 
  #table(d_2014$qpp8_1) #first
  #table(d_2014$qpp8_8) #last

#Original data 2019
d_2019 = read_dta(here("Input", "data", "2019.dta")) %>% 
  clean_names()

  ##ptvs
  #table(d_2019$q10_1) #first
  #table(d_2019$q10_10) #last

```

```{r}
#Extra operations on cumulated data 1989-2004
#select country, year, ptvs
d_1989_2004 = d_1989_2004 %>% 
  select(t_ees, t_cntry, t_var001,
         t_var111, #vote
         t_var112, #vote_choice
         t_var134, #lr
         t_var154, #pol int
         t_var211, #partisan
         t_var212, # party id
         t_var217, #gender
         t_var218, #year of birth
         t_var225, # social class
         t_var_ptv_party1: t_var_ptv_party58) 

# t_var111
d_1989_2004 <- d_1989_2004 %>%
  mutate(
    t_var111 = ifelse(t_var111 == 1, 1, 0) # All values except 1 are recoded as 0
  ) %>%
  rename(vote = t_var111)

# t_var112
d_1989_2004 <- d_1989_2004 %>%
  mutate(t_var112 = ifelse(t_var112 < 100, 0, t_var112)) %>%
  rename(vote_choice = t_var112) 

# t_var134
d_1989_2004 <- d_1989_2004 %>%
  mutate(
    t_var134 = ifelse(t_var134 > 10, NA, t_var134) # Recode values greater than 10 as NA
  ) %>%
  rename(lr = t_var134)

# t_var154
d_1989_2004 <- d_1989_2004 %>%
  mutate(
    t_var154 = ifelse(t_var154 > 7, NA, t_var154), # Recode values greater than 8 as NA
    t_var154 = case_when( # Invert polarity
      t_var154 == 1 ~ 4,
      t_var154 == 2 ~ 3,
      t_var154 == 3 ~ 2,
      t_var154 == 4 ~ 1,
      TRUE ~ t_var154 
    )
  ) %>%
  rename(pol_int = t_var154) # Rename column

#t_var211 #close to a party?
d_1989_2004 <- d_1989_2004 %>%
  mutate(t_var211 = ifelse(t_var211 < 10000, 0, 1)) %>%
  rename(partisan = t_var211) 

# t_var212 # how close?
d_1989_2004 <- d_1989_2004 %>%
  mutate(
    t_var212 = ifelse(t_var212 > 4, 0, t_var212), # Recode values greater than 4 as 0
    t_var212 = case_when( # Invert polarity
      t_var212 == 1 ~ 4,
      t_var212 == 2 ~ 3,
      t_var212 == 3 ~ 2,
      t_var212 == 4 ~ 1,
      TRUE ~ t_var212 # Keep NA values unchanged
    )
  ) %>%
  rename(part_id = t_var212) # Rename column


# t_var217
d_1989_2004 <- d_1989_2004 %>%
  mutate(
    t_var217 = ifelse(t_var217 > 2, NA, t_var217)) %>%
  rename(sex = t_var217)

# t_var218
d_1989_2004 <- d_1989_2004 %>%
  rename(yb = t_var218) %>%
  mutate(yb = ifelse(yb == 0 | yb > 1986, NA, yb)) 

# t_var225
d_1989_2004 <- d_1989_2004 %>%
  mutate(
    t_var225 = ifelse(t_var225 > 5, NA, t_var225)) %>%
  rename(class = t_var225)

# ptvs
d_1989_2004 = d_1989_2004 %>%
  mutate(across(t_var_ptv_party1:t_var_ptv_party58, ~case_when(
    . == 99 ~ NA_real_,
    . == 98 ~ NA_real_,
    . == 97 ~ NA_real_,
    . == 12 ~ NA_real_,
    . == -1 ~ NA_real_,
    TRUE ~ .
  )))

#wave dfs for d_1989_2004
list_of_dfs = split(d_1989_2004, d_1989_2004$t_ees)

for (year in names(list_of_dfs)) {
  var_name = paste("df", year, sep = "_")
  assign(var_name, list_of_dfs[[year]], envir = .GlobalEnv)
}

```

# Processing

## Data managment

### 1989 

```{r}
#split the dataframe into a list of dataframes based on unique values of t_cntry
list_1989 = split(df_1989, df_1989$t_var001)

# remove columns with only NA values
list_1989 = lapply(list_1989, function(df) {
  df %>%
    select(where(~ !all(is.na(.))),
           -c(t_ees, t_cntry, t_var001))
})

#examine pattern of missing
map(list_1989, vis_miss)

#split belgium
list_1989[["29"]] = list_1989[["2"]] %>% 
  select(c(vote:class, t_var_ptv_party8, t_var_ptv_party11, t_var_ptv_party15, t_var_ptv_party20, t_var_ptv_party25, t_var_ptv_party32))

list_1989[["2"]] = list_1989[["2"]] %>% 
  select(-c(t_var_ptv_party8, t_var_ptv_party11, t_var_ptv_party15, t_var_ptv_party20, t_var_ptv_party25, t_var_ptv_party32))

#remove NAs
list_1989 = lapply(list_1989, function(df) {
  df %>% 
    filter(complete.cases(.))
})

#list to Dataframes  
list_1989 = lapply(list_1989, function(x) {as.data.frame(x)})

```

### 1994

```{r}
#split the dataframe into a list of dataframes based on unique values of t_cntry
list_1994 = split(df_1994, df_1994$t_var001)

# remove columns with only NA values
list_1994 = lapply(list_1994, function(df) {
  df %>%
    select(where(~ !all(is.na(.))),
           -c(t_ees, t_cntry, t_var001))
})

#examine pattern of missing
map(list_1994, vis_miss)

#split belgium
list_1994[["29"]] = list_1994[["2"]] %>% 
  select(c(vote:class, t_var_ptv_party9, t_var_ptv_party14, t_var_ptv_party19, t_var_ptv_party24, t_var_ptv_party27, t_var_ptv_party28))

list_1994[["2"]] = list_1994[["2"]] %>% 
  select(-c(t_var_ptv_party9, t_var_ptv_party14, t_var_ptv_party19, t_var_ptv_party24, t_var_ptv_party26, t_var_ptv_party27, t_var_ptv_party28))

#remove NAs
list_1994 = lapply(list_1994, function(df) {
  df %>% 
    filter(complete.cases(.))
})

#list to Dataframes  
list_1994 = lapply(list_1994, function(x) {as.data.frame(x)})

#merge germanies and remove east germany
list_1994[["27"]] = rbind(list_1994[["27"]], list_1994[["28"]])
list_1994[["28"]] = NULL
```

### 1999 

```{r}
#split the dataframe into a list of dataframes based on unique values of t_cntry
list_1999 = split(df_1999, df_1999$t_var001) 

# remove columns with only NA values
list_1999 = lapply(list_1999, function(df) {
  df %>%
    select(where(~ !all(is.na(.))),
           -c(t_ees, t_cntry, t_var001))
})

#examine pattern of missing
map(list_1999, vis_miss)

#split belgium
list_1999[["29"]] = list_1999[["2"]] %>% 
  select(c(vote:class, t_var_ptv_party9, t_var_ptv_party14, t_var_ptv_party19, t_var_ptv_party24, t_var_ptv_party28, t_var_ptv_party29))

list_1999[["2"]] = list_1999[["2"]] %>% 
  select(-c(t_var_ptv_party9, t_var_ptv_party14, t_var_ptv_party19, t_var_ptv_party24, t_var_ptv_party28, t_var_ptv_party29))

#remove NAs
list_1999 = lapply(list_1999, function(df) {
  df %>% 
    filter(complete.cases(.))
})

#list to Dataframes  
list_1999 = lapply(list_1999, function(x) {as.data.frame(x)})

```

### 2004 

```{r}
#here the trend file is wrong. We have ptvs for sweden in the 2004.dta file
SWE_2004 = read_dta(here("Input", "data", "2004.dta")) %>% 
  clean_names() %>% 
  filter(country == 26) %>% 
  select(v111, v112, v134, v154, v211, v212, v217, v218, v224, v115: v128) %>% 
  mutate(across(v115:v128, ~na_if(.x, 99)))

# v111
SWE_2004 <- SWE_2004 %>%
  mutate(
    v111 = ifelse(v111 == 1, 1, 0) # All values except 1 are recoded as 0
  ) %>%
  rename(vote = v111)

# v112
SWE_2004 <- SWE_2004 %>%
  mutate(v112 = ifelse(v112 > 8, 0, v112)) %>%
  rename(vote_choice = v112) 

# v134
SWE_2004 <- SWE_2004 %>%
  mutate(
    v134 = ifelse(v134 > 10, NA, v134) # Recode values greater than 10 as NA
  ) %>%
  rename(lr = v134)

# v154
SWE_2004 <- SWE_2004 %>%
  mutate(
    v154 = ifelse(v154 > 7, NA, v154), # Recode values greater than 8 as NA
    v154 = case_when( # Invert polarity
      v154 == 1 ~ 4,
      v154 == 2 ~ 3,
      v154 == 3 ~ 2,
      v154 == 4 ~ 1,
      TRUE ~ v154 
    )
  ) %>%
  rename(pol_int = v154) # Rename column

#v211 #close to a party?
SWE_2004 <- SWE_2004 %>%
  mutate(v211 = ifelse(v211 < 10000, 0, 1)) %>%
  rename(partisan = v211) 

# v212 # how close?
SWE_2004 <- SWE_2004 %>%
  mutate(
    v212 = ifelse(v212 > 4, 0, v212), # Recode values greater than 4 as 0
    v212 = case_when( # Invert polarity
      v212 == 1 ~ 4,
      v212 == 2 ~ 3,
      v212 == 3 ~ 2,
      v212 == 4 ~ 1,
      TRUE ~ v212 # Keep NA values unchanged
    )
  ) %>%
  rename(part_id = v212) # Rename column


# v217
SWE_2004 <- SWE_2004 %>%
  mutate(
    v217 = ifelse(v217 > 2, NA, v217)) %>%
  rename(sex = v217)

# v218
SWE_2004 <- SWE_2004 %>%
  rename(yb = v218) %>%
  mutate(yb = ifelse(yb == 0 | yb > 1986, NA, yb)) 

# v224
SWE_2004 <- SWE_2004 %>%
  mutate(
    v225 = ifelse(v224 > 5, NA, v224)) %>%
  rename(class = v224)

# ptvs
SWE_2004 = SWE_2004 %>%
  mutate(across(v115:v128, ~case_when(
    . == 99 ~ NA_real_,
    . == 98 ~ NA_real_,
    . == 97 ~ NA_real_,
    . == 12 ~ NA_real_,
    . == -1 ~ NA_real_,
    TRUE ~ .
  )))

# remove columns with only NA values, than NA.omit
SWE_2004 = SWE_2004 %>%
  select(where(~ !all(is.na(.)))) %>% 
  na.omit()

# rename ptv cols
SWE_2004 <- SWE_2004 %>%
  rename_with(~ paste0("t_var_ptv_party", seq_along(.)), .cols = v115:v122)
```


```{r}
# Now back again on the trend file with the same procedure as before

#split the dataframe into a list of dataframes based on unique values of t_cntry
list_2004 = split(df_2004, df_2004$t_var001)

# remove columns with only NA values
list_2004 = lapply(list_2004, function(df) {
  df %>%
    select(where(~ !all(is.na(.))),
           -c(t_ees, t_cntry, t_var001))
})

# remove counrties without PTVs data 
list_2004[["2"]] = NULL
list_2004[["16"]] = NULL
list_2004[["17"]] = NULL

#assign data of SWE
list_2004[["26"]] = SWE_2004

#examine pattern of missing
map(list_2004, vis_miss)

#exclude party19 and party20 from britain
list_2004[["3"]][["t_var_ptv_party19"]] = NULL
list_2004[["3"]][["t_var_ptv_party20"]] = NULL

#remove NAs
list_2004 = lapply(list_2004, function(df) {
  df %>% 
    filter(complete.cases(.))
})

#list to Dataframes  
list_2004 = lapply(list_2004, function(x) {as.data.frame(x)})

```

### 2009 original

```{r}
#here there is a V for splitting Belgium: t103
#select ptvs and country, NAs, recode
d_2009 = d_2009 %>%
  select(t103, q24, q25, q46, q78, q87, q88, q102, q103, q114, q39_p1:q39_p15) %>%
  mutate(across(q39_p1:q39_p15, ~case_when(
    . > 10 ~ NA_real_,
    TRUE ~ .
  )))

# q24
d_2009 <- d_2009 %>%
  mutate(
    q24 = ifelse(q24 == 1, 1, 0) # All values except 1 are recoded as 0
  ) %>%
  rename(vote = q24)

# q25
d_2009 <- d_2009 %>%
  mutate(q25 = ifelse(q25 < 100, 0, q25)) %>%
  rename(vote_choice = q25) 

# q46
d_2009 <- d_2009 %>%
  mutate(
    q46 = ifelse(q46 > 10, NA, q46) # Recode values greater than 10 as NA
  ) %>%
  rename(lr = q46)

# q78
d_2009 <- d_2009 %>%
  mutate(
    q78 = ifelse(q78 > 6, NA, q78), # Recode values greater than 8 as NA
    q78 = case_when( # Invert polarity
      q78 == 1 ~ 4,
      q78 == 2 ~ 3,
      q78 == 3 ~ 2,
      q78 == 4 ~ 1,
      TRUE ~ q78 
    )
  ) %>%
  rename(pol_int = q78) # Rename column

#q87 #close to a party?
d_2009 <- d_2009 %>%
  mutate(q87 = ifelse(q87 < 10000, 0, 1)) %>%
  rename(partisan = q87) 

# q88 # how close?
d_2009 <- d_2009 %>%
  mutate(
    q88 = ifelse(q88 > 4, 0, q88), # Recode values greater than 4 as 0
    q88 = case_when( # Invert polarity
      q88 == 1 ~ 4,
      q88 == 2 ~ 3,
      q88 == 3 ~ 2,
      q88 == 4 ~ 1,
      TRUE ~ q88 # Keep NA values unchanged
    )
  ) %>%
  rename(part_id = q88) # Rename column


# q102
d_2009 <- d_2009 %>%
  mutate(
    q102 = ifelse(q102 > 2, NA, q102)) %>%
  rename(sex = q102)

# q103
d_2009 <- d_2009 %>%
  rename(yb = q103) %>%
  mutate(yb = ifelse(yb == 0 | yb > 1986, NA, yb)) 

# q114
d_2009 <- d_2009 %>%
  mutate(
    q114 = ifelse(q114 > 5, NA, q114)) %>%
  rename(class = q114)

# ptvs
d_2009 <- d_2009 %>%
  mutate(across(q39_p1:q39_p15, ~case_when(
    . == 99 ~ NA_real_,
    . == 98 ~ NA_real_,
    . == 97 ~ NA_real_,
    . == 12 ~ NA_real_,
    . == -1 ~ NA_real_,
    TRUE ~ .
  ))) %>%
  rename_with(~ paste0("t_var_ptv_party", seq_along(.)), .cols = q39_p1:q39_p15)

```


```{r}
#country list
list_2009 = split(d_2009, d_2009$t103)

#remove country and empty col
list_2009 = lapply(list_2009, function(df) {
  df %>%
    select(where(~ !all(is.na(.))),
           -(t103))
})

#examine pattern of missing
map(list_2009, vis_miss)

#remove NAs
list_2009 = lapply(list_2009, function(df) {
  df %>% 
    filter(complete.cases(.))
})

```

### 2014 original

```{r}
#select ptvs and country, NAs, recode
d_2014 = d_2014 %>%
  select(countrycode, p13_intlang, 
         qp1, qp2, qpp13, qp6_9, qpp21, qpp22, d10, vd11, d63,
         qpp8_1:qpp8_8) %>%
  mutate(across(qpp8_1:qpp8_8, ~case_when(
    . <= 0 ~ NA_real_,
    TRUE ~ .
  )))

# qp1
d_2014 <- d_2014 %>%
  mutate(
    qp1 = ifelse(qp1 == 1, 1, 0) # All values except 1 are recoded as 0
  ) %>%
  rename(vote = qp1)

# qp2
d_2014 <- d_2014 %>%
  mutate(qp2 = ifelse(qp2 < 1, 0, qp2)) %>%
  rename(vote_choice = qp2) 

# qpp13
d_2014 <- d_2014 %>%
  mutate(
    qpp13 = ifelse(qpp13 < 1, NA, qpp13) # Recode values greater than 10 as NA
  ) %>%
  rename(lr = qpp13)

# qp6_9
d_2014 <- d_2014 %>%
  mutate(
    qp6_9 = ifelse(qp6_9 < 1, NA, qp6_9), # Recode values greater than 8 as NA
    qp6_9 = case_when( # Invert polarity
      qp6_9 == 1 ~ 4,
      qp6_9 == 2 ~ 3,
      qp6_9 == 3 ~ 2,
      qp6_9 == 4 ~ 1,
      TRUE ~ qp6_9 
    )
  ) %>%
  rename(pol_int = qp6_9) # Rename column

#qpp21 #close to a party?
d_2014 <- d_2014 %>%
  mutate(qpp21 = ifelse(qpp21 < 2, 0, 1)) %>%
  rename(partisan = qpp21) 

# qpp22 # how close?
d_2014 <- d_2014 %>%
  mutate(
    qpp22 = ifelse(qpp22 < 1, 0, qpp22), # Recode values greater than 4 as 0
    qpp22 = case_when( # Invert polarity
      qpp22 == 1 ~ 4,
      qpp22 == 2 ~ 3,
      qpp22 == 3 ~ 2,
      qpp22 == 4 ~ 1,
      TRUE ~ qpp22 # Keep NA values unchanged
    )
  ) %>%
  rename(part_id = qpp22) # Rename column


# d10
d_2014 <- d_2014 %>%
  mutate(
    d10 = ifelse(d10 > 2, NA, d10)) %>%
  rename(sex = d10)

#vd11
d_2014 <- d_2014 %>%
  rename(yb = vd11) %>%
  mutate(yb = 2014 - yb) # Convert age to year of birth

# d63
d_2014 <- d_2014 %>%
  mutate(
    d63 = ifelse(d63 < 1, NA, d63)) %>%
  rename(class = d63)

# ptvs
d_2014 = d_2014 %>%
  mutate(across(qpp8_1:qpp8_8, ~case_when(
    . == -99 ~ NA_real_,
    . == -9 ~ NA_real_,
    . == -8 ~ NA_real_,
    . == -7 ~ NA_real_,
    TRUE ~ .
  )))%>%
  rename_with(~ paste0("t_var_ptv_party", seq_along(.)), .cols = qpp8_1:qpp8_8)


# NEW: rescale all "t_var_ptv_" values: reduce by 1 so that scale is 0 to 10
d_2014 <- d_2014 %>%
  mutate(across(starts_with("t_var_ptv_"), ~ ifelse(!is.na(.), . - 1, NA)))

```


```{r}
#country list
list_2014 = split(d_2014, d_2014$countrycode)

#split belgium
b1 = list_2014[["1056"]] %>% 
  filter(p13_intlang == 4)

b2 = list_2014[["1056"]] %>% 
  filter(p13_intlang == 8) %>% 
  select(-t_var_ptv_party7) #always missing, error documented in the release note

# Assign belgium
list_2014[["1056"]] = b1
list_2014[["1056_2"]] = b2

#remove country, int_lang, and empty col
list_2014 = lapply(list_2014, function(df) {
  df %>%
    select(where(~ !all(is.na(.))),
           -c(countrycode, p13_intlang))
})


#examine pattern of missing
map(list_2014, vis_miss)

#remove NAs
list_2014 = lapply(list_2014, function(df) {
  df %>% 
    filter(complete.cases(.))
})

```

### 2019 original

```{r}
#select ptvs and country, NAs, recode
d_2019 = d_2019 %>%
  select(countrycode, meta_lang_be, 
         q6, q7, q11, q21, q25, q26, d3, d4_1, d7, 
         q10_1:q10_10) 

# q6
d_2019 <- d_2019 %>%
  mutate(
    q6 = ifelse(q6 == 1, 1, 0) # All values except 1 are recoded as 0
  ) %>%
  rename(vote = q6)

# q7
d_2019 <- d_2019 %>%
  mutate(q7 = ifelse(q7 < 100, 0, q7)) %>%
  rename(vote_choice = q7) 

# q11
d_2019 <- d_2019 %>%
  mutate(
    q11 = ifelse(q11 > 10, NA, q11) # Recode values greater than 10 as NA
  ) %>%
  rename(lr = q11)

# q21
d_2019 <- d_2019 %>%
  mutate(
    q21 = ifelse(q21 > 4, NA, q21), # Recode values greater than 8 as NA
    q21 = case_when( # Invert polarity
      q21 == 1 ~ 4,
      q21 == 2 ~ 3,
      q21 == 3 ~ 2,
      q21 == 4 ~ 1,
      TRUE ~ q21 
    )
  ) %>%
  rename(pol_int = q21) # Rename column

#q25 #close to a party?
d_2019 <- d_2019 %>%
  mutate(q25 = ifelse(q25 < 99, 0, 1)) %>%
  rename(partisan = q25) 

# q26 # how close?
d_2019 <- d_2019 %>%
  mutate(
    q26 = ifelse(q26 > 3, 0, q26), # Recode values greater than 4 as 0
    q26 = case_when( # Invert polarity
      q26 == 0 ~ 0,
      q26 == 1 ~ 3,
      q26 == 2 ~ 2,
      q26 == 3 ~ 1,
      TRUE ~ q26 # Keep NA values unchanged
    )
  ) %>%
  rename(part_id = q26) # Rename column


# d3
d_2019 <- d_2019 %>%
  mutate(
    d3 = ifelse(d3 > 2, NA, d3)) %>%
  rename(sex = d3)

# d4_1
d_2019 <- d_2019 %>%
  rename(yb = d4_1) %>%
  mutate(yb = ifelse(yb > 2001, NA, yb)) 

# d7
d_2019 <- d_2019 %>%
  mutate(
    d7 = ifelse(d7 > 5, NA, d7)) %>%
  rename(class = d7)

# ptvs
d_2019 = d_2019 %>%
  mutate(across(q10_1:q10_10, ~case_when(
    . == 99 ~ NA_real_,
    . == 98 ~ NA_real_,
    . == 97 ~ NA_real_,
    . == 96 ~ NA_real_,
    TRUE ~ .
  )))%>%
  rename_with(~ paste0("t_var_ptv_party", seq_along(.)), .cols = q10_1:q10_10)

```

```{r}
#country list
list_2019 = split(d_2019, d_2019$countrycode)

#split belgium
b1 = list_2019[["1056"]] %>% 
  filter(meta_lang_be == 1)

b2 = list_2019[["1056"]] %>% 
  filter(meta_lang_be == 2)

# Assign belgium
list_2019[["1056"]] = b1
list_2019[["1056_2"]] = b2

#remove country and empty col
list_2019 = lapply(list_2019, function(df) {
  df %>%
    select(where(~ !all(is.na(.))),
           -c(countrycode, meta_lang_be))
})

#examine pattern of missing
map(list_2019, vis_miss)

#remove NAs
list_2019 = lapply(list_2019, function(df) {
  df %>% 
    filter(complete.cases(.))
})

#Bulgaria is uncorrectly coded with code 1110, i transform it in 1100 as in 2004 and 2019
names(list_2019)[3] = "1100"
```


## Country codes 

### 1989-2004
```{r}
#conversion rule: labels of this:
#d_1989_2004$t_var001

# Name vector 
name_conversion = c("austria", "belgium", "britain", "cyprus", "czech republic",
                     "denmark", "estonia", "finland", "france", "germany",
                     "greece", "hungary", "ireland", "italy", "latvia",
                     "lithuania", "luxembourg", "malta", "netherlands", "northern ireland",
                     "poland", "portugal", "slovakia", "slovenia", "spain",
                     "sweden", "west germany", "east germany", "belgium")
names(name_conversion) = 1:29

# define function for conversion
rename_elements = function(list) {names(list) = sapply(names(list), 
  function(x) name_conversion[x])
  return(list)}

# Apply the function
list_1989 = rename_elements(list_1989)
list_1994 = rename_elements(list_1994)
list_1999 = rename_elements(list_1999)
list_2004 = rename_elements(list_2004)

#English to ISO
# Convert English country names in list names to ISO codes
names(list_1989) = countrycode(names(list_1989), "country.name", "iso3c")
names(list_1994) = countrycode(names(list_1994), "country.name", "iso3c") 
names(list_1999) = countrycode(names(list_1999), "country.name", "iso3c")
names(list_2004) = countrycode(names(list_2004), "country.name", "iso3c")

# drop northern ireland
list_1989[["NA"]] = NULL
list_1994[["NA"]] = NULL
list_2004[["NA"]] = NULL
```

### 2009

```{r}
#conversion rule from: d_2009$t103

country_codes = c("BEL","BEL","CZE","DNK","DEU","EST","GRC","ESP","FRA","IRL","ITA",
                  "CYP","LVA","LTU","LUX","HUN","MLT","NLD","AUT","POL","PRT","SVN",
                  "SVK","FIN","SWE","GBR","BGR","ROU")

# Assuming list_2009 is your list and it has the same number of elements as the length of country_codes
names(list_2009) <- country_codes

```


### 2014-2019
```{r}
# Create a named vector with numerical codes as names and country names as values
country_codes = setNames(c("Austria", "Belgium", "Bulgaria", "Croatia", "Cyprus", "Czech Rep.", 
                            "Denmark", "Estonia", "Finland", "France", "Germany", "Greece", 
                            "Hungary", "Ireland", "Italy", "Latvia", "Lithuania", "Luxembourg", 
                            "Malta", "Netherlands", "Poland", "Portugal", "Romania", "Slovakia", 
                            "Slovenia", "Spain", "Sweden", "United Kingdom", "Belgium"),
                          c("1040", "1056", "1100", "1191", "1196", "1203", "1208", "1233", "1246",
                            "1250", "1276", "1300", "1348", "1372", "1380", "1428", "1440", "1442",
                            "1470", "1528", "1616", "1620", "1642", "1703", "1705", "1724", "1752", "1826", "1056_2"))

#replace numerical names with country names 
names(list_2014) = country_codes[names(list_2014)]
names(list_2019) = country_codes[names(list_2019)]

#english names to ISO
names(list_2014) = countrycode(names(list_2014), "country.name", "iso3c") 
names(list_2019) = countrycode(names(list_2019), "country.name", "iso3c")

#remove isolate from LUX 2014
list_2014[["LUX"]][["qpp8_6"]] = NULL
```

## Merge

```{r}
#rename belgium2 to avoid common names within each list
names(list_1989)[13] <- "XXX"
names(list_1994)[13] <- "XXX"
names(list_1999)[16] <- "XXX"
names(list_2009)[2] <- "XXX"
names(list_2014)[29] <- "XXX"
names(list_2019)[29] <- "XXX"
```


### Common list

```{r}
EES = list()
EES[["1989"]] = list_1989
EES[["1994"]] = list_1994
EES[["1999"]] = list_1999
EES[["2004"]] = list_2004
EES[["2009"]] = list_2009
EES[["2014"]] = list_2014
EES[["2019"]] = list_2019
```

```{r}
#NEW: remove unplausible PTV data

# After manual checks, we remove this parties from the analyses
table(EES[["2019"]][["POL"]][["t_var_ptv_party8"]])
table(EES[["2019"]][["ROU"]][["t_var_ptv_party9"]])

# remove
EES[["2019"]][["POL"]][["t_var_ptv_party8"]] = NULL
EES[["2019"]][["ROU"]][["t_var_ptv_party9"]] = NULL
```


# Output
```{r}
#save merged file
saveRDS(EES, here("Input", "data", "EES.rds"))
```
